requirements.txt

playwright
python-docx
